requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
distlib==0.3.4
dnspython==2.2.0
email-validator==1.1.3
et-xmlfile==1.1.0
filelock==3.4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
flask-simplelogin==0.1.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.14.1
mysql-connector-python==8.0.27
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.19.2
pytube==11.0.2
pytz==2020.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.15.0
soupsieve==2.3.1
SQLAlchemy==1.3.18
tweepy==4.5.0
urllib3==1.26.8
virtualenv==20.13.0
Werkzeug==1.0.1
WTForms==3.0.1
